backend-server/requirements.txt
Flask
flask-cors
python-dotenv
sqlite3